[9.x] Add a whereOrNull clause
This is purely a syntactical sugar suggestion. We had lots of examples where we've had to recreate this method in code, and have been using this as a macro locally for a while now, and I thought it might be of use in the framework. 

Example use:

An expiry date of a blog entry, where you want to find all blogs that expire after a date or have expiry date set to null.